Adds sleep after dropping connections
Uses partition_point & split_off

Moves tests to separate module
