+++ zebra-network/src/peer_set/initialize/recent_by_ip.rs
@@ -8,6 +8,9 @@ use std::{
 
 use crate::constants;
 
+#[cfg(test)]
+mod tests;
+
 #[derive(Debug)]
 /// Stores IPs of recently attempted inbound connections.
 pub struct RecentByIp {
@@ -50,7 +53,9 @@ impl RecentByIp {
     /// Returns true if there was a recently attempted inbound connection.
     pub fn is_past_limit_or_add(&mut self, ip: IpAddr) -> bool {
         let now = Instant::now();
-        self.prune_by_time(now);
+        if self.by_ip.contains_key(&ip) {
+            self.prune_by_time(now);
+        }
 
         let count = self.by_ip.entry(ip).or_default();
         if *count >= self.max_connections_per_ip {
@@ -64,81 +69,22 @@ impl RecentByIp {
 
     /// Prunes entries older than `time_limit`, decrementing or removing their counts in `by_ip`.
     fn prune_by_time(&mut self, now: Instant) {
-        while let Some((ip, time)) = self.by_time.pop_front() {
-            if now.saturating_duration_since(time) <= self.time_limit {
-                return self.by_time.push_front((ip, time));
-            }
+        // `by_time` must be sorted for this to work.
+        let split_off_idx = self
+            .by_time
+            .partition_point(|&(_, time)| now.saturating_duration_since(time) > self.time_limit);
 
-            if let Some(count) = self.by_ip.get_mut(&ip) {
+        let updated_by_time = self.by_time.split_off(split_off_idx);
+
+        for (ip, _) in &self.by_time {
+            if let Some(count) = self.by_ip.get_mut(ip) {
                 *count -= 1;
                 if *count == 0 {
-                    self.by_ip.remove(&ip);
+                    self.by_ip.remove(ip);
                 }
             }
         }
+
+        self.by_time = updated_by_time;
     }
-}
-
-#[test]
-fn old_connection_attempts_are_pruned() {
-    const TEST_TIME_LIMIT: Duration = Duration::from_secs(5);
-
-    let _init_guard = zebra_test::init();
-
-    let mut recent_connections = RecentByIp::new(Some(TEST_TIME_LIMIT), None);
-    let ip = "127.0.0.1".parse().expect("should parse");
-
-    assert!(
-        !recent_connections.is_past_limit_or_add(ip),
-        "should not be past limit"
-    );
-    assert!(
-        recent_connections.is_past_limit_or_add(ip),
-        "should be past max_connections_per_ip limit"
-    );
-
-    std::thread::sleep(TEST_TIME_LIMIT / 3);
-
-    assert!(
-        recent_connections.is_past_limit_or_add(ip),
-        "should still contain entry after a third of the time limit"
-    );
-
-    std::thread::sleep(3 * TEST_TIME_LIMIT / 4);
-
-    assert!(
-        !recent_connections.is_past_limit_or_add(ip),
-        "should prune entry after 13/12 * time_limit"
-    );
-
-    const TEST_MAX_CONNS_PER_IP: usize = 3;
-
-    let mut recent_connections =
-        RecentByIp::new(Some(TEST_TIME_LIMIT), Some(TEST_MAX_CONNS_PER_IP));
-
-    for _ in 0..TEST_MAX_CONNS_PER_IP {
-        assert!(
-            !recent_connections.is_past_limit_or_add(ip),
-            "should not be past limit"
-        );
-    }
-
-    assert!(
-        recent_connections.is_past_limit_or_add(ip),
-        "should be past max_connections_per_ip limit"
-    );
-
-    std::thread::sleep(TEST_TIME_LIMIT / 3);
-
-    assert!(
-        recent_connections.is_past_limit_or_add(ip),
-        "should still be past limit after a third of the reconnection delay"
-    );
-
-    std::thread::sleep(3 * TEST_TIME_LIMIT / 4);
-
-    assert!(
-        !recent_connections.is_past_limit_or_add(ip),
-        "should prune entry after 13/12 * time_limit"
-    );
 }
